Modules and ES6 syntax
