+++ index.html
@@ -6,6 +6,7 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <link rel="stylesheet" href="./css/index.css">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <script type="module" src="modules/index.js"></script>
   <title>Awesome Books</title>
 </head>
 
@@ -56,7 +57,6 @@
     <footer>
       <span class="span-copyright">Copyright -@2022</span>
     </footer>
-    <script src="app.js"></script>
   </main>
 </body>
